Find the Missing Number In Given array
You are given a list of n-1 integers and these integers are in the range of 1 to n. There are no duplicates in list. One of the integers is missing in the list. Write an efficient code to find the missing integer.
Test case 1
I/P    [1, 2, 4, ,6, 3, 7, 8]
O/P    5
Test case 2
I/P    [1, 5, 4, ,6, 3, 7, 8]
O/P    2